api/requirements.txt
Pillow==10.0.1
pillow-heif==0.13.0
